Buildtools Dockerfile
buildtools/ contains a Dockerfile which can build the buildtools packages. This
is done with Ubuntu 22.04 as it can run bitbake with the tools it gets from the
live package feeds. Images which do not have the appropriate tooling in their
package feeds will be able to copy the appropriate buildtools installer.

This usage has been implemented in the following distros...

- ubuntu-18.04
- leap-15.3
- leap-15.4
- almalinux-8.7

Many tweaks to the `Build images` workflow to allow for this to happen
within github actions.