One platform per job
It seems doing both platforms in the same job takes way longer, possibly
longer than I am able to run a Github action runner, so this is an
attempt to continue using runners by building the separate platforms on
within different jobs. This may be reverted based on the result of the
HEAD~1.